fixing code duplication and ove the identical function to basedvgeo
